fix(coding): corrige regressões da decomposição do QuestionsPanel
Endereça os 3 achados confirmados na revisão do PR #395:
- SortableQuestion volta a usar o componente compartilhado
  FieldHeaderLabel (em vez de reimplementar o bloco de
  prefixo+descrição+help_text à mão), preservando a consistência
  visual com ComparisonPanel e AutoReviewFieldPanel (#373/#365).
- AssignedCodingViewProps volta a exigir submitting/notes/
  onNotesChange/readOnly/onReorder em tempo de compilação, via
  Omit<QuestionsPanelProps, ...> + redeclaração, em vez de herdá-los
  como opcionais de QuestionsPanelProps.
- OutOfScopeConfig passa a ser definido em useOutOfScopeState.ts
  (dono do tipo) e reexportado por QuestionsPanel.tsx, corrigindo a
  dependência invertida do hook para o componente que o chama.
